# vm: Use bn254 gnark-based pairing check