--- a/erigon-lib/crypto/bn256/gnark_bn254.go
+++ b/erigon-lib/crypto/bn256/gnark_bn254.go
@@ -17,13 +17,14 @@
 package bn256
 
 import (
+	"encoding/binary"
 	"errors"
 
 	"github.com/consensys/gnark-crypto/ecc/bn254"
 )
 
-// UnmarshalCurvePoint unmarshals a given input [32-byte X | 32-byte Y] slice to a G1Affine point
-func UnmarshalCurvePoint(input []byte, point *bn254.G1Affine) error {
+// UnmarshalCurvePointG1 unmarshals a given input [32-byte X | 32-byte Y] slice to a G1Affine point
+func UnmarshalCurvePointG1(input []byte, point *bn254.G1Affine) error {
 	if len(input) != 64 {
 		return errors.New("invalid input size")
 	}
@@ -54,11 +55,64 @@ func UnmarshalCurvePoint(input []byte, point *bn254.G1Affine) error {
 	return nil
 }
 
-// MarshalCurvePoint marshals a given G1Affine point to byte slice with [32-byte X | 32-byte Y] form
-func MarshalCurvePoint(point *bn254.G1Affine, ret []byte) []byte {
+// MarshalCurvePointG1 marshals a given G1Affine point to byte slice with [32-byte X | 32-byte Y] form
+func MarshalCurvePointG1(point *bn254.G1Affine, ret []byte) []byte {
 	xBytes := point.X.Bytes()
 	yBytes := point.Y.Bytes()
 	ret = append(ret, xBytes[:]...)
 	ret = append(ret, yBytes[:]...)
+	return ret
+}
+
+// UnmarshalCurvePointG2 unmarshals a given input [64-byte X | 64-byte Y] slice to a G2Affine point
+func UnmarshalCurvePointG2(input []byte, point *bn254.G2Affine) error {
+	if len(input) != 32*4 {
+		return errors.New("invalid input size")
+	}
+
+	isAllZeroes := true
+	for i := 0; i < 32*4; i += 8 {
+		if 0 != binary.BigEndian.Uint64(input[i:i+8]) {
+			isAllZeroes = false
+			break
+		}
+	}
+	if isAllZeroes {
+		return nil
+	}
+
+	// read X and Y coordinates
+	// p.X.A1 | p.X.A0
+	if err := point.X.A1.SetBytesCanonical(input[:32]); err != nil {
+		return err
+	}
+	if err := point.X.A0.SetBytesCanonical(input[32 : 32*2]); err != nil {
+		return err
+	}
+	// p.Y.A1 | p.Y.A0
+	if err := point.Y.A1.SetBytesCanonical(input[32*2 : 32*3]); err != nil {
+		return err
+	}
+	if err := point.Y.A0.SetBytesCanonical(input[32*3 : 32*4]); err != nil {
+		return err
+	}
+
+	// subgroup check
+	if !point.IsInSubGroup() {
+		return errors.New("invalid point: subgroup check failed")
+	}
+	return nil
+}
+
+// MarshalCurvePointG2 marshals a given G2Affine point to byte slice with [64-byte X | 64-byte Y] form
+func MarshalCurvePointG2(point *bn254.G2Affine, ret []byte) []byte {
+	x0Bytes := point.X.A0.Bytes()
+	x1Bytes := point.X.A1.Bytes()
+	y0Bytes := point.Y.A0.Bytes()
+	y1Bytes := point.Y.A1.Bytes()
+	ret = append(ret, x0Bytes[:]...)
+	ret = append(ret, x1Bytes[:]...)
+	ret = append(ret, y0Bytes[:]...)
+	ret = append(ret, y1Bytes[:]...)
 	return ret
 }